modificacion en forma de obtener usuario y roles. Ahora se obtiene desde el layout principal de auth
